+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0
+
+Performance improvment for large files. Only the visible range of the active editor gets decorated.\
+Typing, scrolling (active document only) or changing active editor trigger re-decoration.
+
 ## 1.1.0
 
 ![Duration editing](duration_editing.gif)